moved internal/pool/stats.Stats to internal/pool.Stats
